Normalize username string by replacing diacritics with standard chars
Also needs to be pointed out that for some reason letter ł in PL locale
is not recognized as a diacritic character. Therefore it is needed to
replace it individually.
In this commit also there have been some changes in css. This is the
beginning of making components style-independent (so they will inherit
styling from the whole app. Goal is to leave only css needed for
displaying correctly. And not necessarily beautiful. ;]